pyproject.toml
[tool.poetry]
name = "cv-edu"
version = "0.1.0"
description = ""
authors = ["sameertantry <sameertantry@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = "^1.26.0"
torch = "^2.1.0"
torchvision = "^0.16.0"
matplotlib = "^3.8.0"
pillow = "^10.0.1"
opencv-python = "^4.8.1.78"
seaborn = "^0.13.0"
pandas = "^2.1.1"
omegaconf = "^2.3.0"
albumentations = "^1.3.1"
lightning = "^2.1.0"
fire = "^0.5.0"
dvc = "^3.30.3"
dvc-gdrive = "^2.20.0"
hydra-core = "^1.3.2"
torchmetrics = "^1.2.1"
mlflow = "^2.8.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
pytest = "^7.4.2"

[tool.black]
line-length = 90
target-version = ["py311"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.flake8]
max-line-length = 90

[tool.nbqa.config]
black = "pyproject.toml"
isort = "pyproject.toml"
flake8 = "setup.cfg"

[tool.nbqa.addopts]
flake8 = ["--extend-ignore=E402"]

[tool.nbqa.mutate]
black = 1
isort = 1

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
